+++ Assets/Scripts/Managers/GBAIsometric/GBAIsometric_Spyro_Manager.cs
@@ -314,8 +314,8 @@ namespace R1Engine
 
             var rom = FileFactory.Read<GBAIsometric_Spyro_ROM>(GetROMFilePath, context);
 
-            Func<byte, Unity_MapCollisionTypeGraphic> collGraphicFunc = x => ((GBAIsometric_Spyro3_TileCollisionType2D)x).GetCollisionTypeGraphic();
-            Func<byte, string> collNameFunc = x => ((GBAIsometric_Spyro3_TileCollisionType2D)x).ToString();
+            Func<ushort, Unity_MapCollisionTypeGraphic> collGraphicFunc = x => ((GBAIsometric_Spyro3_TileCollisionType2D)x).GetCollisionTypeGraphic();
+            Func<ushort, string> collNameFunc = x => ((GBAIsometric_Spyro3_TileCollisionType2D)x).ToString();
             if (context.Settings.EngineVersion < EngineVersion.GBAIsometric_Spyro3) {
                 collGraphicFunc = x => ((GBAIsometric_Spyro2_TileCollisionType2D)x).GetCollisionTypeGraphic();
                 collNameFunc = x => ((GBAIsometric_Spyro2_TileCollisionType2D)x).ToString();
